Share code to trigger production and aggregation duties.
